requirements.txt

paho-mqtt